# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API Documentation",
        "description": "API support for a E-Commerce",
        "contact": {
            "email": "makinsuroju@gmail.com"
        },
        "license": {
            "name": "Apache License 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registration",
                "operationId": "c8443fb95bbf4a97f23c39cbaf099216",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstName",
                                    "lastName",
                                    "email",
                                    "password",
                                    "telephone"
                                ],
                                "properties": {
                                    "firstName": {
                                        "title": "firstName",
                                        "type": "string",
                                        "default": "John",
                                        "minLength": 3
                                    },
                                    "lastName": {
                                        "title": "lastName",
                                        "type": "string",
                                        "default": "Doe",
                                        "minLength": 3
                                    },
                                    "email": {
                                        "title": "email",
                                        "type": "email",
                                        "default": "JohnDoe@mail.com"
                                    },
                                    "password": {
                                        "title": "password",
                                        "type": "password",
                                        "default": "********",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "title": "password_confirmation",
                                        "type": "password",
                                        "default": "********",
                                        "minLength": 8
                                    },
                                    "telephone": {
                                        "title": "telephone",
                                        "type": "telephone",
                                        "default": "07067980879"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "title": "success",
                                            "type": "boolean",
                                            "default": "true"
                                        },
                                        "message": {
                                            "title": "message",
                                            "type": "string",
                                            "example": "Created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Request not processed"
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "15400f94bfe16d4952aa545606939fc2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "title": "email",
                                        "type": "email",
                                        "default": "john-doe@mail.com"
                                    },
                                    "password": {
                                        "title": "password",
                                        "type": "password",
                                        "default": "********",
                                        "minLength": 8
                                    },
                                    "remember_me": {
                                        "title": "remember_me",
                                        "type": "boolean",
                                        "default": "false"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "title": "success",
                                            "type": "boolean",
                                            "default": "true"
                                        },
                                        "access_token": {
                                            "title": "access_token",
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "title": "token_type",
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "title": "success",
                                            "type": "boolean",
                                            "default": "false"
                                        },
                                        "message": {
                                            "title": "message",
                                            "type": "string",
                                            "default": "Invalid Credentials!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Request not processed"
                    }
                }
            }
        },
        "/api/v1/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "operationId": "39f13ab593b99e98cae286d8e71e1444",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "title": "success",
                                            "type": "boolean",
                                            "default": "true"
                                        },
                                        "message": {
                                            "title": "message",
                                            "type": "string",
                                            "example": "Successfully logout"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/signup/activate/{token}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "activate account",
                "operationId": "c6bbf91643c10f337156e0d2cddcfbc5",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "unsuccessful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "title": "message",
                                            "type": "string",
                                            "example": "This activation token is invalid"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/activation-token": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Request new activation token",
                "operationId": "fc871a06c40329dd2b4d5a8000088fc9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "title": "email",
                                        "type": "email",
                                        "default": "john-doe@mail.com",
                                        "example": "john@mail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "title": "success",
                                            "type": "boolean",
                                            "default": "true"
                                        },
                                        "message": {
                                            "title": "message",
                                            "type": "string",
                                            "default": "A new verification link has been sent to your email address"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "title": "success",
                                            "type": "boolean",
                                            "default": "false"
                                        },
                                        "message": {
                                            "title": "message",
                                            "type": "string",
                                            "default": "User not found, please register"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "properties": {
                    "id": {
                        "title": "id",
                        "type": "integer",
                        "default": "1"
                    },
                    "first_name": {
                        "title": "first_name",
                        "type": "string",
                        "default": "John",
                        "minLength": 3
                    },
                    "last_name": {
                        "title": "last_name",
                        "type": "string",
                        "default": "Doe",
                        "minLength": 3
                    },
                    "email": {
                        "title": "email",
                        "type": "email",
                        "default": "john-doe@mail.com"
                    },
                    "avatar": {
                        "title": "avatar",
                        "type": "string",
                        "default": "avatar.png"
                    },
                    "telephone": {
                        "title": "telephone",
                        "type": "telephone",
                        "default": "080 2566 4567"
                    },
                    "activation_token": {
                        "title": "activation_token",
                        "type": "string",
                        "example": "6wjKmRtjddXW79sKXsOrJ9ZnloOY5iuwVm9SlA9Le3xSR8ydD2ugljZwNhPZ"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}